Use upcoming version of libdbusmenu-qt
This contains a temporary fork of the importer, with the main paths
fixed.
The old code spawned a new event loop in the signal QMenu::aboutToShow()
This refetched the menu, even though we had already fetched it, before
calling show.

Spawning new event loops from a QML function, leads to all sorts of
crashes. This
fixes that, and saves some pointless DBus traffic

BUG: 343971
BUG: 345838
BUG: 345933
REVIEW: 123992