Preserve the order in dbus menu when doing action reusing.
Summary:
Right now, new Action always appends to the list, but application may
want to add new action in the middle of the menu. This breaks such
application with dynamic menu entries. This change tries to keep the
order of actions same as the order in dbus request.

Test Plan: Manually tested the brokened menu in certain applications.

Reviewers: #plasma, davidedmundson, broulik

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8585