libdbusmenu-qt: Remove nonexistant actions directly from the menu
Summary:
The getLayout response handler compares the new list of actions with the
current actions in the menu and calls deleteLater on all actions which aren't
part of the new list anymore.
Then, it adds all actions from the new list which aren't part of the menu yet.

As deleteLater only has an effect after the next event processing, the menu
still contains them together with the added actions.
This resulted in broken size calculations, as even for static menus the item
count changed during aboutToShow.

Note that this is not a proper solution for the resize issue, as the
aboutToShow handler changes the menus content for a reason, the application
is free to add/remove items at any point in time.

Test Plan:
The context menu for SNI items in the system tray shows at the
correct position now. Only tested the wayland session.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11586